requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
SQLAlchemy==1.4.50
psycopg2-binary==2.9.6
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==1.10.7
python-dotenv==1.0.0
typing-extensions==4.8.0
email-validator==2.1.0.post1
pydantic[email]

